Add W25Q32 Flash memory testing functionality
- Implement complete W25Q32 SPI Flash driver with error handling
- Add comprehensive test suite including quick and full testing modes
- Integrate SPI communication layer with hardware abstraction
- Add printf redirection to UART for test output
- Support basic operations: read, write, erase (page/sector/chip)
- Include power management functions (sleep/wake)
- Add boundary condition testing and error validation
- Replace W24C02 EEPROM tests with W25Q32 Flash tests
- Update main.c to run W25Q32 quick test on startup

Features:
- Professional-grade driver with timeout protection
- Support for cross-page operations with auto-truncation
- Multi-page continuous read/write capabilities
- Performance testing and power management validation
- Clear test result output via UART

+++ Core/Src/w24c02.c
@@ -449,7 +449,7 @@ void register_W24C02_ReadBytes(uint8_t innerAddr, uint8_t *bytes, uint8_t len) {
   // 循环读取多个数据字节
   for (uint8_t i = 0; i < len; i++) {
     /* 读取第i个数据字节 */
-    uint8_t timeout = 0xFFFF; // 为每个字节设置独立的超时计数器
+    uint16_t timeout = 0xFFFF; // 为每个字节设置独立的超时计数器
 
     /* 等待数据接收完成 */
     while (!(I2C2->SR1 & I2C_SR1_RXNE) && timeout) {
